chore: move server code to app router
