+++ next.config.js
@@ -12,6 +12,10 @@ module.exports = {
     includePaths: [path.join(__dirname, 'src/styles')],
     prependData: `@import "common.scss";`,
   },
+  experimental: {
+    // Necessary to prevent https://github.com/sequelize/sequelize/issues/16589
+    serverComponentsExternalPackages: ['sequelize'],
+  },
   async headers() {
     return [
       {
