+++ package.json
@@ -50,7 +50,7 @@
     "react-syntax-highlighter": "^15.5.0",
     "react-use": "^17.5.0",
     "react18-json-view": "^0.2.9-canary.0",
-    "sequelize": "^6.37.1",
+    "sequelize": "^6.37.3",
     "sharp": "^0.33.2",
     "twilio": "^5.0.1"
   },
